fix migration

+++ omega-groups/db/migrate/20111103170906_create_groups_topics.rb
@@ -8,6 +8,6 @@ class CreateGroupsTopics < ActiveRecord::Migration
   end
 
   def self.down
-    drop_table :groups_topics
+    drop_table :groups_topics if table_exists?(:groups_topics)
   end
 end
